long atoi

+++ libft.h
@@ -38,7 +38,7 @@ char				*ft_strrchr(const char *s, int c);
 char				*ft_strnstr(const char *big, const char *little,
 						size_t len);
 signed int			ft_strncmp(const char *s1, const char *s2, size_t n);
-int					ft_atoi(const char *ptr);
+long				ft_atoi(const char *ptr);
 int					ft_isalpha(int c);
 int					ft_isdigit(int c);
 int					ft_isalnum(int c);
